Add core dependencies and update project structure

[2025-08-26 19:14:19]
