+++ package.json
@@ -53,8 +53,12 @@
     "react": "^18.2.0",
     "react-day-picker": "^8.9.1",
     "react-dom": "^18.2.0",
+    "react-error-boundary": "^6.0.0",
     "react-helmet-async": "^2.0.5",
     "react-hook-form": "^7.47.0",
+    "react-hot-toast": "^2.6.0",
+    "react-i18next": "^15.7.2",
+    "react-redux": "^9.2.0",
     "react-remove-scroll-bar": "^2.3.8",
     "react-router-dom": "^6.20.0",
     "recharts": "^2.8.0",
