+++ src/components/FeaturedListingsSection.jsx
@@ -1,0 +1,407 @@
+import React, { useState } from 'react';
+import { motion, AnimatePresence } from 'framer-motion';
+import { Star, Eye, Heart, ArrowRight, Clock, Users, Award } from 'lucide-react';
+import { Link } from 'react-router-dom';
+
+const featuredListings = [
+    {
+        id: 1,
+        title: "AI-Powered E-commerce Platform",
+        category: "Web Development",
+        description: "Modern e-commerce solution with AI-driven product recommendations and personalized shopping experience",
+        rating: 4.9,
+        reviews: 127,
+        views: 2847,
+        likes: 156,
+        image: "https://images.unsplash.com/photo-1551434678-e076c223a692?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2070&q=80",
+        price: "$15,000",
+        tags: ["React", "Node.js", "AI/ML", "E-commerce"],
+        duration: "3-4 months",
+        team: "5 experts",
+        location: "Remote",
+        featured: true,
+        technologies: ["React", "Node.js", "MongoDB", "AI/ML", "AWS"],
+        highlights: ["AI Recommendations", "Real-time Analytics", "Mobile Responsive", "SEO Optimized"]
+    },
+    {
+        id: 2,
+        title: "Mobile Banking Application",
+        category: "Mobile Development",
+        description: "Secure and user-friendly mobile banking app with biometric authentication and real-time transactions",
+        rating: 4.8,
+        reviews: 89,
+        views: 1956,
+        likes: 134,
+        image: "https://images.unsplash.com/photo-1563013544-824ae1b704d3?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2070&q=80",
+        price: "$22,000",
+        tags: ["React Native", "Security", "FinTech", "Biometrics"],
+        duration: "4-5 months",
+        team: "6 experts",
+        location: "Hybrid",
+        featured: true,
+        technologies: ["React Native", "Node.js", "PostgreSQL", "Biometrics", "Security"],
+        highlights: ["Biometric Auth", "Real-time Transactions", "Security Compliance", "Cross-platform"]
+    },
+    {
+        id: 3,
+        title: "Cloud Infrastructure Migration",
+        category: "Cloud & DevOps",
+        description: "Complete migration from on-premise to cloud with automated CI/CD pipelines and monitoring",
+        rating: 4.7,
+        reviews: 156,
+        views: 3241,
+        likes: 189,
+        image: "https://images.unsplash.com/photo-1451187580459-43490279c0fa?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2072&q=80",
+        price: "$18,500",
+        tags: ["AWS", "Docker", "Kubernetes", "CI/CD"],
+        duration: "2-3 months",
+        team: "4 experts",
+        location: "On-site",
+        featured: true,
+        technologies: ["AWS", "Docker", "Kubernetes", "Jenkins", "Prometheus"],
+        highlights: ["Zero Downtime", "Auto-scaling", "Monitoring", "Security"]
+    },
+    {
+        id: 4,
+        title: "Healthcare AI Diagnostic System",
+        category: "AI & Machine Learning",
+        description: "Advanced medical diagnostic platform using computer vision and machine learning for accurate disease detection",
+        rating: 4.9,
+        reviews: 203,
+        views: 4567,
+        likes: 278,
+        image: "https://images.unsplash.com/photo-1576091160399-112ba8d25d1f?ixlib=rb-4.0.3&auto=format&fit=crop&w=2070&q=80",
+        price: "$35,000",
+        tags: ["Python", "TensorFlow", "Computer Vision", "Healthcare"],
+        duration: "6-8 months",
+        team: "8 experts",
+        location: "Remote",
+        featured: true,
+        technologies: ["Python", "TensorFlow", "OpenCV", "Docker", "AWS"],
+        highlights: ["95% Accuracy", "Real-time Processing", "HIPAA Compliant", "API Integration"]
+    },
+    {
+        id: 5,
+        title: "Smart City IoT Platform",
+        category: "IoT Solutions",
+        description: "Comprehensive IoT platform for smart city management with real-time monitoring and analytics",
+        rating: 4.6,
+        reviews: 78,
+        views: 1890,
+        likes: 112,
+        image: "https://images.unsplash.com/photo-1449824913935-59a10b8d2000?ixlib=rb-4.0.3&auto=format&fit=crop&w=2070&q=80",
+        price: "$28,000",
+        tags: ["IoT", "Big Data", "Analytics", "Smart Cities"],
+        duration: "5-6 months",
+        team: "7 experts",
+        location: "Hybrid",
+        featured: false,
+        technologies: ["IoT Sensors", "Apache Kafka", "Elasticsearch", "React", "Node.js"],
+        highlights: ["Real-time Monitoring", "Predictive Analytics", "Scalable Architecture", "Dashboard"]
+    }
+];
+
+const specialServices = [
+    {
+        title: "AI Consulting",
+        description: "Strategic AI implementation guidance",
+        icon: <Award className="w-5 h-5"/>,
+        link: "/services/ai-consulting"
+    },
+    {
+        title: "Cloud Migration",
+        description: "Seamless cloud infrastructure transition",
+        icon: <Award className="w-5 h-5"/>,
+        link: "/services/cloud-migration"
+    },
+    {
+        title: "Cybersecurity",
+        description: "Enterprise-grade security solutions",
+        icon: <Award className="w-5 h-5"/>,
+        link: "/services/cybersecurity"
+    },
+    {
+        title: "Digital Transformation",
+        description: "End-to-end business transformation",
+        icon: <Award className="w-5 h-5"/>,
+        link: "/services/digital-transformation"
+    }
+];
+
+export function FeaturedListingsSection() {
+    const [hoveredListing, setHoveredListing] = useState(null);
+    
+    const containerVariants = {
+        hidden: { opacity: 0 },
+        visible: {
+            opacity: 1,
+            transition: {
+                staggerChildren: 0.1,
+                delayChildren: 0.2
+            }
+        }
+    };
+    
+    const itemVariants = {
+        hidden: { y: 30, opacity: 0 },
+        visible: {
+            y: 0,
+            opacity: 1,
+            transition: {
+                duration: 0.6,
+                ease: "easeOut"
+            }
+        }
+    };
+    
+    const cardVariants = {
+        hidden: { scale: 0.9, opacity: 0 },
+        visible: {
+            scale: 1,
+            opacity: 1,
+            transition: {
+                duration: 0.4,
+                ease: "easeOut"
+            }
+        }
+    };
+    
+    const renderStars = (rating) => {
+        return Array.from({ length: 5 }, (_, i) => (
+            <span key={i} className={i < rating ? 'text-yellow-400' : 'text-gray-400'}>
+                ★
+            </span>
+        ));
+    };
+    
+    return (
+        <section className="py-20 bg-gradient-to-br from-zion-slate-dark via-zion-blue-dark to-zion-blue relative overflow-hidden">
+            {/* Background decorative elements */}
+            <div className="absolute inset-0 opacity-5">
+                <div className="absolute top-20 left-20 w-32 h-32 border border-zion-cyan rounded-full"></div>
+                <div className="absolute bottom-20 right-20 w-24 h-24 border border-zion-purple rounded-full"></div>
+                <div className="absolute top-1/2 left-1/2 w-16 h-16 border border-zion-cyan-light rounded-full"></div>
+            </div>
+            
+            <div className="container mx-auto px-4 relative z-10">
+                <motion.div 
+                    className="text-center mb-16" 
+                    initial={{ opacity: 0, y: 20 }} 
+                    whileInView={{ opacity: 1, y: 0 }} 
+                    viewport={{ once: true }} 
+                    transition={{ duration: 0.6 }}
+                >
+                    <h2 className="text-4xl md:text-5xl font-bold text-white mb-6">
+                        Featured <span className="bg-gradient-to-r from-zion-cyan to-zion-purple bg-clip-text text-transparent">Projects</span>
+                    </h2>
+                    <p className="text-zion-slate-light text-lg md:text-xl max-w-3xl mx-auto leading-relaxed">
+                        Discover our most successful and innovative projects. Each one represents our commitment to excellence 
+                        and our ability to deliver cutting-edge solutions.
+                    </p>
+                </motion.div>
+                
+                <motion.div 
+                    className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8 mb-16"
+                    variants={containerVariants}
+                    initial="hidden"
+                    whileInView="visible"
+                    viewport={{ once: true }}
+                >
+                    {featuredListings.map((listing, index) => (
+                        <motion.div 
+                            key={listing.id} 
+                            variants={itemVariants}
+                            onHoverStart={() => setHoveredListing(listing.id)}
+                            onHoverEnd={() => setHoveredListing(null)}
+                            whileHover={{ y: -8 }}
+                            transition={{ type: "spring", stiffness: 300 }}
+                        >
+                            <div className="bg-gradient-to-br from-zion-blue-dark/80 to-zion-blue-dark/40 backdrop-blur-sm border border-zion-blue-light/30 rounded-2xl overflow-hidden hover:border-zion-cyan/50 transition-all duration-300 hover:shadow-2xl hover:shadow-zion-cyan/20 group h-full">
+                                {/* Image */}
+                                <div className="relative h-48 overflow-hidden">
+                                    <img 
+                                        src={listing.image} 
+                                        alt={listing.title}
+                                        className="w-full h-full object-cover group-hover:scale-110 transition-transform duration-500"
+                                    />
+                                    {listing.featured && (
+                                        <div className="absolute top-4 right-4 bg-gradient-to-r from-zion-cyan to-zion-purple text-white px-3 py-1 rounded-full text-xs font-semibold">
+                                            Featured
+                                        </div>
+                                    )}
+                                    <div className="absolute top-4 left-4 bg-zion-slate-dark/80 backdrop-blur-sm text-white px-3 py-1 rounded-full text-xs">
+                                        {listing.category}
+                                    </div>
+                                </div>
+                                
+                                {/* Content */}
+                                <div className="p-6">
+                                    <h3 className="text-xl font-bold text-white mb-3 group-hover:text-zion-cyan transition-colors">
+                                        {listing.title}
+                                    </h3>
+                                    <p className="text-zion-slate-light text-sm mb-4 line-clamp-2">
+                                        {listing.description}
+                                    </p>
+                                    
+                                    {/* Duration and team */}
+                                    <div className="grid grid-cols-2 gap-3 mb-4">
+                                        <div className="flex items-center gap-2 text-zion-slate-light/80 text-xs">
+                                            <Clock className="w-3 h-3"/>
+                                            <span>{listing.duration}</span>
+                                        </div>
+                                        <div className="flex items-center gap-2 text-zion-slate-light/80 text-xs">
+                                            <Users className="w-3 h-3"/>
+                                            <span>{listing.team}</span>
+                                        </div>
+                                    </div>
+                                    
+                                    {/* Tags */}
+                                    <div className="flex flex-wrap gap-2 mb-4">
+                                        {listing.tags.slice(0, 3).map((tag, index) => (
+                                            <span key={index} className="px-2 py-1 rounded-md bg-zion-slate/30 text-zion-slate-light text-xs">
+                                                {tag}
+                                            </span>
+                                        ))}
+                                        {listing.tags.length > 3 && (
+                                            <span className="text-zion-cyan/60 text-xs">
+                                                +{listing.tags.length - 3} more
+                                            </span>
+                                        )}
+                                    </div>
+                                    
+                                    {/* Stats */}
+                                    <div className="flex items-center justify-between mb-4 text-sm text-zion-slate-light">
+                                        <div className="flex items-center gap-1">
+                                            <Star className="w-4 h-4 text-yellow-400 fill-current"/>
+                                            <span>{listing.rating}</span>
+                                            <span>({listing.reviews})</span>
+                                        </div>
+                                        <div className="flex items-center gap-4">
+                                            <div className="flex items-center gap-1">
+                                                <Eye className="w-4 h-4"/>
+                                                <span>{listing.views}</span>
+                                            </div>
+                                            <div className="flex items-center gap-1">
+                                                <Heart className="w-4 h-4"/>
+                                                <span>{listing.likes}</span>
+                                            </div>
+                                        </div>
+                                    </div>
+                                    
+                                    {/* Price and CTA */}
+                                    <div className="flex items-center justify-between mt-auto">
+                                        <span className="text-2xl font-bold text-zion-cyan">
+                                            {listing.price}
+                                        </span>
+                                        <button className="inline-flex items-center gap-2 bg-gradient-to-r from-zion-purple to-zion-purple-dark hover:from-zion-purple-light hover:to-zion-purple text-white px-4 py-2 rounded-lg font-medium transition-all duration-300 transform hover:scale-105">
+                                            View Details
+                                            <ArrowRight className="w-4 h-4"/>
+                                        </button>
+                                    </div>
+
+                                    {/* Expanded details on hover */}
+                                    <AnimatePresence>
+                                        {hoveredListing === listing.id && (
+                                            <motion.div 
+                                                className="mt-4 p-4 rounded-xl bg-zion-blue-dark/60 backdrop-blur-sm border border-zion-cyan/30" 
+                                                initial={{ opacity: 0, height: 0, y: 10 }} 
+                                                animate={{ opacity: 1, height: "auto", y: 0 }} 
+                                                exit={{ opacity: 0, height: 0, y: 10 }} 
+                                                transition={{ duration: 0.3 }}
+                                            >
+                                                <h4 className="text-zion-cyan font-semibold text-sm mb-3">Key Highlights:</h4>
+                                                <div className="space-y-2 mb-4">
+                                                    {listing.highlights.map((highlight, idx) => (
+                                                        <motion.div 
+                                                            key={idx} 
+                                                            className="flex items-center gap-2 text-zion-slate-light/80 text-xs" 
+                                                            initial={{ opacity: 0, x: -10 }} 
+                                                            animate={{ opacity: 1, x: 0 }} 
+                                                            transition={{ delay: idx * 0.1 }}
+                                                        >
+                                                            <div className="w-2 h-2 bg-zion-cyan rounded-full"></div>
+                                                            <span>{highlight}</span>
+                                                        </motion.div>
+                                                    ))}
+                                                </div>
+
+                                                <h4 className="text-zion-cyan font-semibold text-sm mb-3">Technologies:</h4>
+                                                <div className="flex flex-wrap gap-2">
+                                                    {listing.technologies.slice(0, 4).map((tech, idx) => (
+                                                        <motion.span 
+                                                            key={idx} 
+                                                            className="px-2 py-1 bg-zion-blue-light/20 text-zion-cyan text-xs rounded-full border border-zion-cyan/30" 
+                                                            initial={{ opacity: 0, scale: 0.8 }} 
+                                                            animate={{ opacity: 1, scale: 1 }} 
+                                                            transition={{ delay: idx * 0.1 }}
+                                                        >
+                                                            {tech}
+                                                        </motion.span>
+                                                    ))}
+                                                </div>
+                                            </motion.div>
+                                        )}
+                                    </AnimatePresence>
+                                </div>
+                            </div>
+                        </motion.div>
+                    ))}
+                </motion.div>
+                
+                {/* Enhanced featured services section */}
+                <motion.div 
+                    className="mt-16" 
+                    initial={{ opacity: 0, y: 20 }} 
+                    whileInView={{ opacity: 1, y: 0 }} 
+                    viewport={{ once: true }} 
+                    transition={{ duration: 0.6, delay: 0.3 }}
+                >
+                    <h3 className="text-center text-2xl font-bold text-white mb-8">Featured Services</h3>
+                    <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-4 max-w-4xl mx-auto">
+                        {specialServices.map((service, index) => (
+                            <motion.div 
+                                key={service.title} 
+                                initial={{ opacity: 0, y: 20 }} 
+                                whileInView={{ opacity: 1, y: 0 }} 
+                                viewport={{ once: true }} 
+                                transition={{ duration: 0.4, delay: index * 0.1 }} 
+                                whileHover={{ y: -4 }}
+                            >
+                                <Link to={service.link} className="group block">
+                                    <div className="px-6 py-4 bg-zion-blue-light/20 hover:bg-zion-blue-light/30 border border-zion-purple/20 hover:border-zion-purple/50 rounded-xl text-zion-cyan transition-all duration-300 backdrop-blur-sm hover:shadow-lg hover:shadow-zion-purple/20">
+                                        <div className="flex items-center gap-3 mb-2">
+                                            <div className="text-zion-purple-light group-hover:text-zion-purple transition-colors">
+                                                {service.icon}
+                                            </div>
+                                            <span className="font-semibold text-sm">{service.title}</span>
+                                        </div>
+                                        <p className="text-zion-slate-light/80 text-xs leading-relaxed">
+                                            {service.description}
+                                        </p>
+                                    </div>
+                                </Link>
+                            </motion.div>
+                        ))}
+                    </div>
+                    
+                    {/* View all services button */}
+                    <motion.div 
+                        className="text-center mt-8" 
+                        initial={{ opacity: 0, scale: 0.9 }} 
+                        whileInView={{ opacity: 1, scale: 1 }} 
+                        viewport={{ once: true }} 
+                        transition={{ duration: 0.5, delay: 0.5 }}
+                    >
+                        <Link 
+                            to="/comprehensive-services" 
+                            className="inline-flex items-center gap-3 px-8 py-4 bg-gradient-to-r from-zion-purple to-zion-purple-dark hover:from-zion-purple-light hover:to-zion-purple border border-zion-purple/50 hover:border-zion-purple rounded-xl text-white transition-all duration-300 shadow-lg hover:shadow-zion-purple/25 font-semibold group"
+                        >
+                            View All Services
+                            <ArrowRight className="w-5 h-5 group-hover:translate-x-1 transition-transform duration-300"/>
+                        </Link>
+                    </motion.div>
+                </motion.div>
+            </div>
+        </section>
+    );
+}
