+++ src/components/auth/login/index.jsx
@@ -1,0 +1,2 @@
+export { LoginForm as LoginContent } from './LoginForm';
+export { default as LoginErrorFallback } from './LoginErrorFallback';
