+++ src/components/ui/badge.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-const Badge = ({ children, className = '', variant = 'default' }) => {
+export const Badge = ({ children, className = '', variant = 'default' }) => {
     const baseClasses = 'inline-flex items-center rounded-full border px-2.5 py-0.5 text-xs font-semibold transition-colors';
     const variantClasses = {
         default: 'border-transparent bg-primary text-primary-foreground',
@@ -7,7 +7,9 @@ const Badge = ({ children, className = '', variant = 'default' }) => {
         destructive: 'border-transparent bg-destructive text-destructive-foreground',
         outline: 'text-foreground'
     };
-    return (<div className={`${baseClasses} ${variantClasses[variant]} ${className}`}>
-      {children}
-    </div>);
+    return (
+        <div className={`${baseClasses} ${variantClasses[variant]} ${className}`}>
+            {children}
+        </div>
+    );
 };
