--- src/layout/MainNavigation.js
@@ -1,1 +1,0 @@
-export {};
