+++ src/layout/MainNavigation.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+export function MainNavigation() {
+    return (
+        <nav className="hidden md:flex space-x-8">
+            <Link to="/services" className="text-gray-700 hover:text-zion-purple transition-colors">
+                Services
+            </Link>
+            <Link to="/solutions" className="text-gray-700 hover:text-zion-purple transition-colors">
+                Solutions
+            </Link>
+            <Link to="/about" className="text-gray-700 hover:text-zion-purple transition-colors">
+                About
+            </Link>
+            <Link to="/contact" className="text-gray-700 hover:text-zion-purple transition-colors">
+                Contact
+            </Link>
+        </nav>
+    );
+}
